requirements.txt
pandas
plotly_express
steamlit